+++ app/src/lib/photos-library/model/asset.ts
@@ -1,8 +1,12 @@
 import path from 'path';
 import {AssetID} from '../../icloud/icloud-photos/query-parser.js';
 import {FileType} from './file-type.js';
-import {Stats} from 'fs';
+import { Stats } from 'fs';
+import fs from 'fs/promises'
 import {PEntity} from './photos-entity.js';
+import { iCPSApp } from '../../../app/icloud-app.js';
+import { iCPSError } from '../../../app/error/error.js';
+import { LIBRARY_ERR } from '../../../app/error/error-codes.js';
 
 /**
  * Representing the possible asset types
@@ -203,22 +207,32 @@ export class Asset implements PEntity<Asset> {
 
     /**
      * Verifies that the object representation matches the given file
-     * @param file - The read file
-     * @param fileStats - The file stats object to investigate the metadata
+     * @param filePath - The path, where the file is expected
      * @returns True if the provided file matches this object representation
+     * @throws An error, if verification fails
      */
-    verify(file: Buffer, fileStats: Stats): boolean {
-        return this.verifyChecksum(file) && this.verifySize(file) && this.verifyMTime(fileStats);
-    }
+    async verify(filePath: string): Promise<boolean> {
+        let fileStat: Stats
+        try {
+            fileStat = await fs.stat(filePath)
+        } catch (err) {
+            throw new iCPSError(LIBRARY_ERR.ASSET_NOT_FOUND)
+                .addCause(err)
+                .addMessage(filePath)
+        }
 
-    /**
-     * Verifies that the modified timestamps matches the one of the given file
-     * (due to potential rounding errors, we are checking if the timestamp is within 10ms)
-     * @param fileStats - The file stats object to investigate the metadata
-     * @returns True if the modified time matches
-     */
-    private verifyMTime(fileStats: Stats): boolean {
-        return this.withinRange(fileStats.mtimeMs, this.modified, 10);
+        if(!this.withinRange(fileStat.mtimeMs, this.modified, 1000)) {
+            throw new iCPSError(LIBRARY_ERR.ASSET_MODIFICATION_TIME)
+                .addMessage(`${filePath} modification time ${fileStat.mtimeMs}, iCloud ${this.modified}`)
+                .addContext('out-of-range', fileStat.mtimeMs - this.modified)
+        }
+
+        if(fileStat.size !== this.size) {
+            throw new iCPSError(LIBRARY_ERR.ASSET_SIZE)
+                .addMessage(`${filePath} size ${fileStat.size}, iCloud ${this.size}`)
+        }
+
+        return true
     }
 
     /**
@@ -229,17 +243,8 @@ export class Asset implements PEntity<Asset> {
      * @returns true if within range, false otherwise
      */
     private withinRange(x: number, y: number, range: number): boolean {
-        return x > y - range
-            && x < y + range;
-    }
-
-    /**
-     * Compares the size of the file on disk with the size in this object
-     * @param file - The read file
-     * @returns True if the size matches
-     */
-    private verifySize(file: Buffer): boolean {
-        return file.byteLength === this.size;
+        return x >= y - range
+            && x <= y + range;
     }
 
     /**
@@ -248,9 +253,9 @@ export class Asset implements PEntity<Asset> {
      * @param file - The read file
      * @returns True if checksum matches
      */
+    /*
     private verifyChecksum(file: Buffer): boolean {
         return file !== undefined;
-        /*
         Const hashes = [
             `BLAKE2b512`,
             `BLAKE2s256`,
@@ -319,8 +324,8 @@ export class Asset implements PEntity<Asset> {
                     }
                 }
             });
-        }); */
-    }
+        }); 
+    }*/
 
     /**
      *