+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   * Filetype support added for
     * `avi`
   * Reworked error structure
+  * Removed 2GB limitation during verification (closes [#155](https://github.com/steilerDev/icloud-photos-sync/issues/155))
+  * Improves output when asset verification fails (closes [#156](https://github.com/steilerDev/icloud-photos-sync/issues/156))
+  * Acceptable mtime range (before verification fails) increased to 1000ms
 
 ## v1.0.1